requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
sqlparse==0.5.3
drf-spectacular==0.29.0
Pillow==12.0.0